ui fix, added event handler for ui-viewModel
