--- .idea/detekt.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="DetektPluginSettings">
-    <option name="enableFormatting" value="true" />
-  </component>
-</project>
