--- app/src/main/java/com/example/effective_labs/App.kt
@@ -1,7 +1,0 @@
-package com.example.effective_labs
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class App : Application()
