--- app/src/main/java/com/example/effective_labs/ui/HeroDataUi.kt
@@ -1,7 +1,0 @@
-package com.example.effective_labs.ui
-
-data class HeroDataUi(
-    val name: String,
-    val description: String,
-    val imageId: Int,
-)
