--- app/src/main/java/com/example/effective_labs/ui/consts/Dimens.kt
@@ -1,11 +1,0 @@
-package com.example.effective_labs.ui.consts
-
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.ui.unit.dp
-
-val heroesListColumnPaddingLandscape = PaddingValues(top = 8.dp, bottom = 16.dp)
-val heroesListColumnPaddingPortrait = PaddingValues(top = 16.dp, bottom = 60.dp)
-val heroesListImagePaddingLandscape = PaddingValues(top = 8.dp, bottom = 8.dp)
-val heroesListImagePaddingPortrait = PaddingValues(top = 36.dp, bottom = 16.dp)
-val heroesListItemViewPadding = PaddingValues(16.dp)
-val heroesListView = PaddingValues(bottom = 16.dp)
