+++ app/src/main/kotlin/ru/flobsterable/effective_labs/App.kt
@@ -1,0 +1,7 @@
+package ru.flobsterable.effective_labs
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class App : Application()
