+++ app/src/main/kotlin/ru/flobsterable/effective_labs/ui/model/HeroDataUi.kt
@@ -1,0 +1,7 @@
+package ru.flobsterable.effective_labs.ui.model
+
+data class HeroDataUi(
+    val name: String,
+    val description: String,
+    val imageId: Int,
+)
