+++ app/src/main/kotlin/ru/flobsterable/effective_labs/ui/screens/heroesList/model/HeroesListEvent.kt
@@ -1,0 +1,5 @@
+package ru.flobsterable.effective_labs.ui.screens.heroesList.model
+
+sealed class HeroesListEvent {
+    object LoadHeroesList : HeroesListEvent()
+}
