+++ app/src/main/kotlin/ru/flobsterable/effective_labs/ui/screens/heroesList/model/HeroesListViewState.kt
@@ -1,0 +1,7 @@
+package ru.flobsterable.effective_labs.ui.screens.heroesList.model
+
+import ru.flobsterable.effective_labs.ui.model.HeroDataUi
+
+data class HeroesListViewState(
+    val heroesList: List<HeroDataUi> = emptyList()
+)
