+++ app/src/main/kotlin/ru/flobsterable/effective_labs/ui/screens/view/TransparentSystemBars.kt
@@ -1,0 +1,25 @@
+package ru.flobsterable.effective_labs.ui.screens.view
+
+import androidx.compose.material.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.compositeOver
+import com.google.accompanist.systemuicontroller.rememberSystemUiController
+import ru.flobsterable.effective_labs.ui.theme.BlackScrim
+
+@Composable
+fun TransparentSystemBars() {
+    val systemUiController = rememberSystemUiController()
+    val useDarkIcons = MaterialTheme.colors.isLight
+    SideEffect {
+        systemUiController.setSystemBarsColor(
+            color = Color.Transparent,
+            darkIcons = useDarkIcons,
+            isNavigationBarContrastEnforced = false,
+            transformColorForLightContent = { original ->
+                BlackScrim.compositeOver(original)
+            }
+        )
+    }
+}
