+++ app/src/main/kotlin/ru/flobsterable/effective_labs/ui/theme/Color.kt
@@ -1,0 +1,16 @@
+package ru.flobsterable.effective_labs.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Brown = Color(0xFFCC9B6D)
+val Beige = Color(0xFFF1CA89)
+val LightBeige = Color(0xFFF2DAC3)
+val Gray = Color(0xFFC8C2BC)
+
+val darkPrimary = Color(0xFF261C2C)
+val darkSecondary = Color(0xFF3E2C41)
+val darkTertiary = Color(0xFF5C527F)
+val darkCC = Color(0xFF6E85B2)
+val Pink40 = Color(0xFF7D5260)
+
+val BlackScrim = Color(0f, 0f, 0f, 0.3f)
