+++ app/src/main/kotlin/ru/flobsterable/effective_labs/ui/theme/Theme.kt
@@ -1,0 +1,53 @@
+package ru.flobsterable.effective_labs.ui.theme
+
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.darkColors
+import androidx.compose.material.lightColors
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Color.Companion.White
+
+private val DarkColorPalette = darkColors(
+    primary = darkPrimary,
+    secondary = darkSecondary,
+    secondaryVariant = darkTertiary,
+    background = darkCC,
+    onSecondary = White
+)
+
+private val LightColorPalette = lightColors(
+    primary = Brown,
+    secondary = Beige,
+    secondaryVariant = LightBeige,
+    background = Pink40,
+    onSecondary = Color.Black
+
+    /* Other default colors to override
+    background = Color.White,
+    surface = Color.White,
+    onPrimary = Color.White,
+    onSecondary = Color.Black,
+    onBackground = Color.Black,
+    onSurface = Color.Black,
+    */
+)
+
+@Composable
+fun EffectiveLabsTheme(
+    darkTheme: Boolean = isSystemInDarkTheme(),
+    content: @Composable () -> Unit
+) {
+    val colors = if (darkTheme) {
+        DarkColorPalette
+    } else {
+        LightColorPalette
+    }
+
+    MaterialTheme(
+        colors = colors,
+        typography = Typography,
+        shapes = Shapes,
+        content = content
+    )
+}
