+++ app/src/main/kotlin/ru/flobsterable/effective_labs/utils/EventHandler.kt
@@ -1,0 +1,5 @@
+package ru.flobsterable.effective_labs.utils
+
+interface EventHandler<E> {
+    fun obtainEvent(event: E)
+}
